pyproject.toml
[tool.poetry]
name = "wgengine"
version = "0.1.0"
description = ""
authors = ["yux <yuWorm@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.21.5"
aerich = "^0.7.2"
sanic = "^24.6.0"
sanic-ext = "^23.12.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
aiomysql = "^0.2.0"
cryptography = "^43.0.0"
click = "^8.1.7"
jinja2 = "^3.1.4"
sanic-session = "^0.8.0"
redis = "^5.0.8"
loguru = "^0.7.2"
typeguard = "^4.3.0"
pycryptodome = "^3.20.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"


[tool.aerich]
tortoise_orm = "database.conf.db_conf"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
